feat/duplicates: allow duplicates to be present
Sometimes we find that different deps A, B, C depend on different versions of same dep Z. So A depends on Z-ver-x, B on Z-ver-y and C on Z-ver-z for e.g. In such a case we will have duplicate names for rlibs but we don't want to prune to just one as they are all actually different. One solution is to read the Cargo.lock file to see how many versions of a dep are present and leave around that many duplicates. This does not work because in some cases cargo clippy for instance produces its own set of Z's which also we must not prune. So we are going to keep all versions of a lib around if they don't differ by more than a const number of secs as specified in the code. This will help prevent recompilation of such deps even if nothing was upgraded (because in the approach with pruning all duplicates we will have pruned all currently needed versions too).